backend/requirements.txt
django==2.1.7
uwsgi==2.0.18
psycopg2==2.7.7
djangorestframework==3.9.2


djangorestframework
django-cors-headers
markdown
html2text
bleach
beautifulsoup4
Pillow
requests
